pdf_ai_webapp/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1
python-multipart>=0.0.9
pymupdf>=1.24
pillow>=10.0
openai>=1.30
pydantic-settings>=2.2
python-dotenv>=1.0
